add logout func
